msrit-scraper/requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
psycopg2-binary>=2.9.9
cryptography>=41.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
httpx>=0.27.0
